==> ipynb/05-benchmark-scoring.ipynb <==
# Benchmark Scoring

####  Goals
* Obtain a benchmark score using 4 different models.

#### Output Scores
* LassoCV: 0.964363
* BayesianRidge: 0.972271	
* DecisionTree: 0.913885	
* KNN: 0.861510


cd ..

%run lib/__init__.py
%matplotlib inline

X_test = pd.read_pickle('./data/X_test_dummies_df.pkl')
X_train = pd.read_pickle('./data/X_train_dummies_df.pkl')

y_train = pd.read_pickle('./data/y_train.pkl')
y_test = pd.read_pickle('./data/y_test.pkl')

X_train.shape, y_train.shape

## 1. Benchmark Models

# data is already deskewed and scaled.
def benchmark_score(model, model_name, X_train, X_test, y_train, y_test):

    model       = model()
    model.fit(X_train, y_train)
    train_score = model.score(X_train, y_train)
    test_score  = model.score(X_test, y_test)
    
    output = []
    output.append({'model name'  : model_name,
                  'train score' : train_score,
                  'test score'  : test_score})
    return output

results_bm = []

### LassoCV

lasso_bm = benchmark_score(LassoCV, 'LassoCV (bm)', X_train, X_test, y_train, y_test)
results_bm += lasso_bm

### Bayesian Ridge Regressor

bayesian_ridge_bm = benchmark_score(BayesianRidge, 'BayesianRidge (bm)', X_train, X_test, y_train, y_test)
results_bm += bayesian_ridge_bm

### Decision Tree Regressor

decision_tree_bm = benchmark_score(DecisionTreeRegressor, 'DecisionTree (bm)', X_train, X_test, y_train, y_test)
results_bm += decision_tree_bm

### K Neighbors Regressor

knn_bm = benchmark_score(KNeighborsRegressor, 'KNN (bm)', X_train, X_test, y_train, y_test)
results_bm += knn_bm

## 2. Display Benchmark Scores

results_bm_df = pd.DataFrame(results_bm)
results_bm_df

## 3. Pickling

results_bm_df.to_pickle('./data/bm_results_df.pkl')
