==> ipynb/06-remove-outliers.ipynb <==
# Remove Outliers

####  Goals
* Deskew and Scale the entire DataFrame together
* Use Tukey's Method to define outliers
* Remove outliers from encoded dataset and from dataset.

#### Output
* removed outliers from all datasets.

cd ..

%run lib/__init__.py
%matplotlib inline

# the whole dataset. numeric and categorical together.
commute_df = pd.read_pickle('./data/dropped_correlated_features_df.pkl')
commute_df.shape

commute_stats_df = pd.read_pickle('./data/commute_stats_dropped_correlated_features_df.pkl')
commute_stats_df.shape

## 1. Identify Numerical Features in `commute_df`
* Make a DataFrame only containing numerical features

numerical_columns = list(commute_stats_df.index)
len(numerical_columns)

numeric_df = commute_df[numerical_columns]
numeric_df.shape

# add an amount to everything so that boxcox will work.
numeric_df += 1E-9

## 2. Deskew and Scale Numeric Features
* On numeric features only.
* MinMaxScaler first to ensure values are all positive.

from lib.preprocessing import BoxCoxTransformer

pipeline = Pipeline([
    ('boxcox'  , BoxCoxTransformer()),
    ('ss'      , StandardScaler())
])

numeric_deskewed_scaled = pipeline.fit_transform(numeric_df)

numeric_deskewed_scaled_df = pd.DataFrame(numeric_deskewed_scaled, 
                                          columns=numeric_df.columns,
                                          index=numeric_df.index)
numeric_deskewed_scaled_df.head()

## 3. Encode Categorical
* Delete UUID columns, do not want to predict on UUID
* Encode `commute_df`, then put deskewed/scaled numeric features into the dataset in appropriate columns.

commute_df.drop(['UUID'], axis=1, inplace=True)

commute_dummies = pd.get_dummies(commute_df)
commute_dummies.head()

for col in numeric_deskewed_scaled_df.columns:
    commute_dummies[col] = numeric_deskewed_scaled_df[col]

commute_dummies.head()

## 4. Outliers
* Identify outliers in the dataset using Tukey's Method.
* Count outliers in each feature.
* Remove outliers that appear in 4 or more features (this number was chosen as the result represents about 2% of the data).

numeric_deskewed_scaled_df.shape

def display_outliers(dataframe, col, tukey_value):
    Q1 = np.percentile(dataframe[col], 25)
    Q3 = np.percentile(dataframe[col], 75)
    tukey_window = tukey_value*(Q3-Q1)
    less_than_Q1 = dataframe[col] < (Q1 - tukey_window)
    greater_than_Q3 = dataframe[col] > (Q3 + tukey_window)
    tukey_mask = (less_than_Q1 | greater_than_Q3)
    return dataframe[tukey_mask]

outliers = []
for col in numeric_deskewed_scaled_df:
    outlier_df = display_outliers(numeric_deskewed_scaled_df, col, 1.70)
    print(col, outlier_df.shape)
    outliers.append((col, outlier_df))

from collections import Counter
outlier_indices = [list(out[1].index) for out in outliers]
outlier_indices = [outlier_index for sublist in outlier_indices for outlier_index in sublist]

cn = Counter(outlier_indices)

outlier_counts = pd.DataFrame([{'id': key, 'count': val} for key, val in cn.items()])
outlier_counts.set_index('id', inplace=True)
outlier_counts.sort_values('count', ascending=False, inplace=True)
outlier_counts[outlier_counts['count'] > 4].count()

(31/1038)*100

outliers = outlier_counts[outlier_counts['count'] > 4]
outliers

outlier_indices = list(outliers.index)

removed_outliers_full_df = commute_dummies.drop(outlier_indices)
removed_outliers_full_df.shape

numeric_removed_outliers_df = numeric_deskewed_scaled_df.drop(outlier_indices)
numeric_removed_outliers_df.shape

removed_outliers_raw_df = commute_df.drop(outlier_indices)
removed_outliers_raw_df.shape

## 5. Pickling and Saving

# full dataframe, deskewed and scaled, with removed outliers.
removed_outliers_full_df.to_pickle('./data/removed_outliers_full_df.pkl')

# numeric dataframe, deskewed and scaled, with removed outliers
numeric_removed_outliers_df.to_pickle('./data/numeric_removed_outliers_df.pkl')

# raw dataframe with removed outliers to deskew and scale again.
removed_outliers_raw_df.to_pickle('./data/removed_outliers_raw_df.pkl')
